Ember 1.13 compatibility
The `registerComponentLookup` initializer is now an instance initializer, so we need to change the ember-islands boot initializer to an instance initializer too. Unfortunately there's no clean way to plumb the `bypass` config flag through to the instance, as instances do not have access to the parent application. To get around this we just set the config flag directly on the default/deprecated instance that exists in 1.13.

Note that these changes mean that ember-islands will only work in Ember 1.13+.